# public/apiDocsV1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "fedora-backend",
    "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
    "contact": {
      "email": ""
    },
    "version": "1.0.0",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/api/v1",
...
  "schemes": ["http", "https"],
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded",
    "multipart/form-data"
  ],
  "produces": ["application/json"],
  "paths": {
    "/user/register": {
      "post": {
        "deprecated": false,
        "tags": ["User"],
        "description": "Login",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "accountNumber",
            "in": "formData",
            "description": "accountNumber of your bank account",
            "required": true,
            "type": "integer"
          },
          {
            "name": "membershipId",
            "in": "formData",
            "description": "membershipId",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "name",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "email",
           
            "type": "string"
          },
          {
            "name": "mobileNo",
            "in": "formData",
            "description": "your registered mobileNo",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": ["Username: ", "Password: "]
          }
        ]
      }
    },
    "/user/userDetails": {
      "get": {
        "deprecated": false,
        "tags": ["User"],
        "description": "Login",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "customerId",
            "in": "query",
            "description": "userId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": ["Username: ", "Password: "]
          }
        ]
      }
    },
    "/user/getUserBeneficiary": {
      "get": {
        "deprecated": false,
        "tags": ["Beneficiary"],
        "description": "Login",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "customerId",
            "in": "query",
            "description": "userId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": ["Username: ", "Password: "]
          }
        ]
      }
    },
    "/user/deleteBeneficiary": {
      "delete": {
        "deprecated": false,
        "tags": ["Beneficiary"],
        "description": "deleteBeneficiary",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "beneficiaryId",
            "in": "query",
            "description": "beneficiaryId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": ["Username: ", "Password: "]
          }
        ]
      }
    },
    "/user/logout": {
      "delete": {
        "deprecated": false,
        "tags": ["User"],
        "description": "Login",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": ["Username: ", "Password: "]
          }
        ]
      }
    },
    "/user/forgotPassword": {
      "put": {
        "deprecated": false,
        "tags": ["Admin"],
        "description": "Login",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "description": "email",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": ["Username: ", "Password: "]
          }
        ]
      }
    },
    "/user/resetPassword": {
      "put": {
        "deprecated": false,
        "tags": ["Admin"],
        "description": "Login",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "email",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": ["Username: ", "Password: "]
          }
        ]
      }
    },
    "/user/addUser": {
      "post": {
        "deprecated": false,
        "tags": ["Admin"],
        "description": "Login",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "membershipId",
            "in": "formData",
            "description": "membershipId should be capital",
            "required": false,
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "name",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "email",
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "Password",
            "type": "string"
          },
          
          {
            "name": "mobileNo",
            "in": "formData",
            "description": "your registered mobileNo",
            "required": true,
            "type": "integer"
          },
          {
            "name": "allotedStocks",
            "in": "formData",
            "description": "allotedStocks  to customer",
            "required": false,
            "type": "integer"
          },
          {
            "name": "userType",
            "in": "formData",
            "required": true,
            "description": "userType 1 -SuperAdmin  , 2 - normal User by default it will be 2  3-Admin 4-Manager 5 viewer",
            "type": "integer"
          },
          {
            "name": "addressLine1",
            "in": "formData",
            "required": true,
            "description": "addressLine1",
            "type": "string"
          },
          {
            "name": "addressLine2",
            "in": "formData",
            "required": true,
            "description": "addressLine2",
            "type": "string"
          },
          {
            "name": "city",
            "in": "formData",
            "required": true,
            "description": "city",
            "type": "string"
          },
          {
            "name": "state",
            "in": "formData",
            "required": true,
            "description": "state",
            "type": "string"
          },
          {
            "name": "zipCode",
            "in": "formData",
            "required": true,
            "description": "zipCode",
            "type": "integer"
          },
          {
            "name": "country",
            "in": "formData",
            "required": true,
            "description": "country",
            "type": "string"
          },
          {
            "name": "dob",
            "in": "formData",
            "required": true,
            "description": "DOB timeStamp ",
            "type": "string"
          },
          {
            "name": "gender",
            "in": "formData",
            "required": true,
            "description": "gender 1 Male, 2 Female, 3 Others ",
            "type": "integer"
          }
          
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": ["Username: ", "Password: "]
          }
        ]
      }
    },
    "/user/addOtherBankDetails": {
      "post": {
        "deprecated": false,
        "tags": ["Admin"],
        "description": "Login",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "customerId",
            "in": "formData",
            "description": "customerId",
            "required": true,
            "type": "string"
          },
          {
            "name": "accountNumber",
            "in": "formData",
            "description": "accountNumber",
            "required": true,
            "type": "string"
          },
          {
            "name": "IFSCcode",
            "in": "formData",
            "description": "IFSCcode",
            "required": true,
            "type": "string"
          },
          {
            "name": "bankName",
            "in": "formData",
            "description": "bankName",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": ["Username: ", "Password: "]
          }
        ]
      }
    },
    "/user/deleteOtherBankDetails": {
      "delete": {
        "deprecated": false,
        "tags": ["Admin"],
        "description": "Login",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "customerId",
            "in": "formData",
            "description": "customerId",
            "required": true,
            "type": "string"
          },
          {
            "name": "otherBankDetailsId",
            "in": "formData",
            "description": "otherBankDetailsId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "basicAuth": ["Username: ", "Password: "]
          }
        ]
      }
    },
    "/user/updateDeviceToken": {
      "put": {
        "deprecated": false,
        "tags": ["User"],
        "description": "Login",
        "consumes": ["application/x-www-form-urlencoded"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "access token of user",
            "required": true,
            "type": "string"
          },
          {
            "name": "deviceToken",
            "in": "formData",
            "description": "deviceToken",
            "required": false,
            "type": "string"
          }
        ],
      "responses": {
        "200": {
          "description": "Successfully"
        },
        "401": {
          "description": "Unauthorized"
        }
      },
      "security": [
        {
          "basicAuth": ["Username: ", "Password: "]
        }
      ]
    }
  },
...
  }
...
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    }
  }
}